+++ src/GlobalPayments.Api/Terminals/UPA/Responses/UPAResponseHandler.cs
@@ -109,6 +109,7 @@ namespace GlobalPayments.Api.Terminals.UPA.Responses {
         public int? Cvv { get; set; }
         public ThreeDesDukpt ThreeDesDukpt { get; set; }
         public string EmvProcess { get; set; }
+        public bool SafTransaction { get; set; }
 
         //EMV
         public string ApplicationIdentifier { get; set; }